Make the logback appender exit cleanly
The logback appender was set as a daemon thread, which means the JVM
does give an eff if it has cleaned up correctly when it was shutting
down.  This causes the messages that were queued to get sent to just get
dropped.

This is problematic because the messages that are still in the queue are
usually the error messages from shutting down.

This makes the apps in charge of shutting down the thread then, using
the method described here:
http://logback.qos.ch/manual/configuration.html#stopContext

I was looking to see if this was something that should be normally
done...looking at the logback logstash appender under the TCP appenders
section, it suggests shutting down logback cleanly too to ensure all
messages are sent:
https://github.com/logstash/logstash-logback-encoder#tcp-appenders.
Perhaps not a clear precedence, but at least mentioned in another
context